refactor: DTO 수정 및 패키지 수정
