+++ src/main/java/wooteco/subway/domain/path/FindDijkstraShortestPathStrategy.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain.strategy;
+package wooteco.subway.domain.path;
 
 import java.util.List;
 import java.util.Optional;
@@ -7,10 +7,8 @@ import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.WeightedMultigraph;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Path;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
-import wooteco.subway.domain.ShortestPathEdge;
 import wooteco.subway.domain.Station;
 import wooteco.subway.exception.NotFoundException;
