+++ src/main/java/wooteco/subway/domain/path/FindPathStrategy.java
@@ -1,6 +1,5 @@
-package wooteco.subway.domain.strategy;
+package wooteco.subway.domain.path;
 
-import wooteco.subway.domain.Path;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
