+++ src/main/java/wooteco/subway/dto/line/LineSaveRequest.java
@@ -3,6 +3,7 @@ package wooteco.subway.dto.line;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
+import javax.validation.constraints.PositiveOrZero;
 import wooteco.subway.domain.Line;
 
 public class LineSaveRequest {
@@ -22,7 +23,7 @@ public class LineSaveRequest {
     @Positive(message = "상행-하행 노선 길이는 양수 값만 들어올 수 있습니다.")
     private int distance;
 
-    @Min(value = 0, message = "요금의 최소 금액은 0입니다.")
+    @PositiveOrZero(message = "요금의 최소 금액은 0입니다.")
     private int extraFare;
 
     private LineSaveRequest() {
