+++ src/main/java/wooteco/subway/dto/line/LineUpdateRequest.java
@@ -2,6 +2,7 @@ package wooteco.subway.dto.line;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Positive;
+import javax.validation.constraints.PositiveOrZero;
 import wooteco.subway.domain.Line;
 
 public class LineUpdateRequest {
@@ -12,7 +13,7 @@ public class LineUpdateRequest {
     @NotBlank(message = "line 색상은 공백 혹은 null이 들어올 수 없습니다.")
     private String color;
 
-    @Positive(message = "추가 요금은 양수 값만 들어올 수 있습니다.")
+    @PositiveOrZero(message = "요금의 최소 금액은 0입니다.")
     private int extraFare;
 
     private LineUpdateRequest() {
