+++ src/main/java/wooteco/subway/dto/path/PathFindResponse.java
@@ -2,7 +2,7 @@ package wooteco.subway.dto.path;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import wooteco.subway.domain.Path;
+import wooteco.subway.domain.path.Path;
 import wooteco.subway.dto.station.StationResponse;
 
 public class PathFindResponse {
@@ -20,12 +20,12 @@ public class PathFindResponse {
         this.fare = fare;
     }
 
-    public static PathFindResponse from(final Path path, final int age) {
+    public static PathFindResponse from(final Path path, final int fare) {
         List<StationResponse> stationResponses = path.getStations()
                 .stream()
                 .map(StationResponse::from)
                 .collect(Collectors.toList());
-        return new PathFindResponse(stationResponses, path.getDistance(), (int) path.calculateFinalFare(age));
+        return new PathFindResponse(stationResponses, path.getDistance(), fare);
     }
 
     public List<StationResponse> getStations() {
