+++ src/test/java/wooteco/subway/domain/strategy/FindDijkstraShortestPathStrategyTest.java
@@ -8,7 +8,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Path;
+import wooteco.subway.domain.path.FindDijkstraShortestPathStrategy;
+import wooteco.subway.domain.path.FindPathStrategy;
+import wooteco.subway.domain.path.Path;
 import wooteco.subway.domain.Section;
 import wooteco.subway.domain.Sections;
 import wooteco.subway.domain.Station;
